+++ Client/src/Pages/Monitors/monitors.css
@@ -1,4 +1,5 @@
 .current-monitors {
+  margin-top: var(--env-var-spacing-2);
   border: 1px solid #eaecf0;
   padding: 40px;
   border-radius: 4px;
